two components in component folder: DetailDialog, PopUpInfo
