+++ package.json
@@ -6,7 +6,8 @@
   "scripts": {
     "build": "webpack",
     "start": "webpack-dev-server --host 0.0.0.0 --port 8080 --config ./webpack.config.js --mode development",
-    "gencsv": "node data/process_211.js data/211sample.json 211data.csv"
+    "gencsv": "node data/process_211.js data/211sample.json 211data.csv",
+    "test": "jest"
   },
   "author": "Open San Diego",
   "license": "ISC",
@@ -33,8 +34,16 @@
     "@material-ui/icons": "^4.11.2",
     "axios": "^0.21.1",
     "babel-loader": "^8.2.2",
-    "csv": "^5.3.2"
+    "csv": "^5.3.2",
+    "enzyme": "^3.11.0",
+    "enzyme-adapter-react-16": "^1.15.6",
+    "jest": "^26.6.3",
+    "jest-css-modules": "^2.1.0"
   },
+  "jest": {
+    "moduleNameMapper": {
+      "\\.(css|less|scss|sss|styl)$": "<rootDir>/node_modules/jest-css-modules"
+    }},
   "babel": {
     "presets": [
       "@babel/preset-env",
